Remove containers/text, move fixtures to containers/cpstext
